+++ OpenSim/Examples/ExampleMain/OutputReference/TugOfWar_Complete.cpp
@@ -314,12 +314,9 @@ int main()
         ForceReporter* reporter = new ForceReporter(&osimModel);
         osimModel.addAnalysis(reporter);
 
-        // Create the integrator for integrating system dynamics
-        SimTK::RungeKuttaMersonIntegrator integrator(osimModel.getMultibodySystem());
-        integrator.setAccuracy(1.0e-6);
-        
         // Create the manager managing the forward integration and its outputs
-        Manager manager(osimModel,  integrator);
+        Manager manager(osimModel);
+        manager.setIntegratorAccuracy(1.0e-6);
 
         // Print out details of the model
         osimModel.printDetailedInfo(si, cout);